+++ src/bika/lims/utils/analysisrequest.py
@@ -422,11 +422,9 @@ def create_partition(analysis_request, request, analyses, sample_type=None,
     partition.setDateReceived(ar.getDateReceived())
     partition.reindexObject(idxs="getDateReceived")
 
-    # Force partition to same status as the primary
-    status = api.get_workflow_status_of(ar)
-    changeWorkflowState(partition, SAMPLE_WORKFLOW, status)
-    if IReceived.providedBy(ar):
-        alsoProvides(partition, IReceived)
+    # Always set partition to received state
+    changeWorkflowState(partition, SAMPLE_WORKFLOW, "sample_received")
+    alsoProvides(partition, IReceived)
 
     # And initialize the analyses the partition contains. This is required
     # here because the transition "initialize" of analyses rely on a guard,